Revert "#19 enviar notificação de solicitação de entrega para o entregador mais próximo (frontend)"

+++ chegaja_frontend/lib/models/deliveryman/deliveryman.dart
@@ -4,7 +4,7 @@ class Deliveryman {
   String? telefone;
   String? nome;
   String? cpf;
-  double? pesoMaximo;
+  int? pesoMaximo;
   int? id;
 
   Deliveryman({
